obsolete PHP 5.5, test with PHP 7.2
PHP 5.5 is not supported by upstream anymore and hence removed from the
list of supported versions.

As I already use PHP 7.2 for some internal projects, we might run tests
of CDN Linker with it as well. Please keep in mind 7.2 is still a RC,
RC6 as of the time I am writing this.